# crates/codchi-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Codchi API",
    "description": "Per-user codchi-server v1 contract. See v1/phases/00-contract-decisions.md.",
    "version": "1"
  },
  "paths": {
...
    "/v1/machines/{id}": {
      "get": {
        "summary": "Get a machine",
        "operationId": "get_machine",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "default": {
            "description": "Typed API error (ApiError)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Get a machine",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a machine (job)",
        "operationId": "delete_machine",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "default": {
            "description": "Typed API error (ApiError)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Delete a machine (job)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView_for_Null"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "description": "Typed API error. Serialized as an internally-tagged object with a stable `code` discriminator, e.g. `{\"code\":\"machine_not_found\",\"machine\":\"foo\"}`.",
        "oneOf": [
          {
            "description": "No machine with this id exists.",
            "type": "object",
            "required": [
              "code",
              "machine"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "machine_not_found"
                ]
              },
              "machine": {
                "type": "string"
              }
            }
          },
          {
            "description": "Another writer job holds this machine (P7 rule 2).",
            "type": "object",
            "required": [
              "code",
              "job",
              "machine"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "machine_busy"
                ]
              },
              "job": {
                "$ref": "#/components/schemas/JobId"
              },
              "machine": {
                "type": "string"
              }
            }
          },
          {
            "description": "No job with this id exists, or its metadata was pruned.",
            "type": "object",
            "required": [
              "code",
              "job"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "job_not_found"
                ]
              },
              "job": {
                "$ref": "#/components/schemas/JobId"
              }
            }
          },
          {
            "description": "The job is past the point where cancellation is accepted (P7).",
            "type": "object",
            "required": [
              "code",
              "job",
              "state"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "job_not_cancellable"
                ]
              },
              "job": {
                "$ref": "#/components/schemas/JobId"
              },
              "state": {
                "$ref": "#/components/schemas/JobState"
              }
            }
          },
          {
            "description": "The store is unavailable, migrating, or needs repair (P7 rule 5).",
            "type": "object",
            "required": [
              "code",
              "reason"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "store_unavailable"
                ]
              },
              "reason": {
                "type": "string"
              }
            }
          },
          {
            "description": "A store repair/migration job is already running (P7).",
            "type": "object",
            "required": [
              "code",
              "job"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "store_busy"
                ]
              },
              "job": {
                "$ref": "#/components/schemas/JobId"
              }
            }
          },
          {
            "description": "The on-disk SQLite schema must be migrated before serving.",
            "type": "object",
            "required": [
              "code",
              "current",
              "required"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "schema_migration_required"
                ]
              },
              "current": {
                "type": "integer",
                "format": "uint32",
                "minimum": 0.0
              },
              "required": {
                "type": "integer",
                "format": "uint32",
                "minimum": 0.0
              }
            }
          },
          {
            "description": "The client and server disagree on the API major version (Q5).",
            "type": "object",
            "required": [
              "client",
              "code",
              "server"
            ],
            "properties": {
              "client": {
                "type": "integer",
                "format": "uint32",
                "minimum": 0.0
              },
              "code": {
                "type": "string",
                "enum": [
                  "api_version_mismatch"
                ]
              },
              "server": {
                "type": "integer",
                "format": "uint32",
                "minimum": 0.0
              }
            }
          },
          {
            "description": "Implicit start blocked because declared secrets are unset (R5).",
            "type": "object",
            "required": [
              "code",
              "keys",
              "machine"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "missing_required_secrets"
                ]
              },
              "keys": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SecretKey"
                }
              },
              "machine": {
                "type": "string"
              }
            }
          },
          {
            "description": "A request field failed validation (P4). `field` is the request field path (e.g. `id`, `source`, `target`).",
            "type": "object",
            "required": [
              "code",
              "field",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "validation"
                ]
              },
              "field": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          },
          {
            "description": "Unexpected server-side failure.",
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "internal"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Component": {
        "description": "The subsystem a finding pertains to.",
        "type": "string",
        "enum": [
          "server",
          "store",
          "machine",
          "job",
          "migration"
        ]
      },
...
      "EventSeq": {
        "description": "Per-job monotonic event sequence number (Q2). Wire-stable `u64`.",
        "type": "integer",
        "format": "uint64",
        "minimum": 0.0
      },
...
      "Finding": {
        "description": "A single health finding produced by the reconciler or a `doctor_scan`.",
        "type": "object",
        "required": [
          "auto_fixable",
          "code",
          "component",
          "created_at",
          "id",
          "message",
          "severity"
        ],
        "properties": {
          "auto_fixable": {
            "description": "`true` only if repair cannot touch user data or rewrite `flake.lock`.",
            "type": "boolean"
          },
          "code": {
            "description": "Stable code, e.g. `\"store.rootfs_missing\"`.",
            "type": "string"
          },
          "component": {
            "$ref": "#/components/schemas/Component"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/FindingId"
          },
          "machine": {
            "default": null,
            "type": "string",
            "nullable": true
          },
          "message": {
            "description": "User-facing message; may evolve.",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "source_job": {
            "description": "`Some` for `doctor_scan` findings; `None` for background-reconcile.",
            "default": null,
            "$ref": "#/components/schemas/JobId",
            "nullable": true
          },
          "suggested_action": {
            "default": null,
            "type": "string",
            "nullable": true
          }
        }
      },
      "FindingId": {
        "description": "Finding identifier (UUID v7).",
        "type": "string",
        "format": "uuid"
      },
...
      "GenerationId": {
        "description": "Per-machine monotonic generation number.",
        "type": "integer",
        "format": "uint64",
        "minimum": 0.0
      },
      "GenerationStatus": {
        "description": "Lifecycle of a generation. Shared by machine and store generations.",
        "oneOf": [
          {
            "description": "Currently activated.",
            "type": "string",
            "enum": [
              "active"
            ]
          },
          {
            "description": "Built and retained, not active.",
            "type": "string",
            "enum": [
              "inactive"
            ]
          },
          {
            "description": "Pinned against garbage collection.",
            "type": "string",
            "enum": [
              "protected"
            ]
          },
          {
            "description": "Tombstoned; retained only for history.",
            "type": "string",
            "enum": [
              "deleted"
            ]
          }
        ]
      },
      "GenerationView": {
        "description": "One machine generation (a built + activatable system).",
        "type": "object",
        "required": [
          "created_at",
          "flake_lock_hash",
          "id",
          "source_job",
          "status",
          "system_store_path"
        ],
        "properties": {
          "activated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "flake_lock_hash": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/GenerationId"
          },
          "source_job": {
            "description": "The job that produced this generation.",
            "$ref": "#/components/schemas/JobId"
          },
          "status": {
            "$ref": "#/components/schemas/GenerationStatus"
          },
          "system_store_path": {
            "type": "string"
          }
        }
      },
...
      "JobId": {
        "description": "Job identifier (UUID v7).",
        "type": "string",
        "format": "uuid"
      },
      "JobKind": {
        "description": "What a job does. No `Start`/`Stop`/`Restart` (R6); no `Reconcile` (P6).",
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "init",
              "clone",
              "rebuild",
              "update",
              "delete",
              "activate_generation",
              "garbage_collect",
              "migration",
              "doctor_scan",
              "doctor_fix"
            ]
          },
          {
            "description": "R3: read-only flake fetch+eval → [`crate::dto::config::ConfigResolution`].",
            "type": "string",
            "enum": [
              "resolve"
            ]
          },
          {
            "description": "R7: ensure store+machine running, session, env → [`ExecPlan`].",
            "type": "string",
            "enum": [
              "prepare_exec"
            ]
          },
          {
            "description": "R11: bring up the store container (a `Store`-subject job).",
            "type": "string",
            "enum": [
              "store_start"
            ]
          },
          {
            "description": "R11: bounded store recovery (a `Store`-subject job).",
            "type": "string",
            "enum": [
              "store_recover"
            ]
          }
        ]
      },
...
      "JobState": {
        "description": "Job lifecycle states (P7).",
        "type": "string",
        "enum": [
          "queued",
          "running",
          "cancel_requested",
          "cleaning_up",
          "succeeded",
          "failed",
          "cancelled"
        ]
      },
...
      "JobView_for_Null": {
        "description": "The canonical view of a job, generic over its typed success output `O`.\n\nA terminal job is exactly one of: `Failed` with `error: Some(_)`, or `Succeeded` with `output` set — or `output: None` for kinds with no output (R1). The type parameter makes the success payload visible at the call site: kind-specific methods return e.g. `JobView<Rebuilt>` or `JobView<()>` (no output), while the kind-erased [`crate::service::CodchiService::get_job`] returns the default `JobView` = `JobView<JobOutput>`.",
        "type": "object",
        "required": [
          "created_at",
          "id",
          "kind",
          "last_event_seq",
          "state",
          "subject"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "error": {
            "description": "`Some` iff terminal `Failed` (R1/R2).",
            "default": null,
            "$ref": "#/components/schemas/ApiError",
            "nullable": true
          },
          "finished_at": {
            "default": null,
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "id": {
            "$ref": "#/components/schemas/JobId"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "last_event_seq": {
            "description": "Highest event sequence emitted so far.",
            "$ref": "#/components/schemas/EventSeq"
          },
          "output": {
            "description": "`Some` iff terminal `Succeeded` for kinds that produce output (R1/R2).",
            "type": "null",
            "nullable": true
          },
          "started_at": {
            "default": null,
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "subject": {
            "description": "The log source this job acts on (R11). Machine-less operations (`resolve_config`, `doctor_*`, `migration`) are `Server`-subject; store operations are `Store`-subject.",
            "$ref": "#/components/schemas/LogSource"
          }
        }
      },
...
      "LogSource": {
        "description": "A long-lived log-emitting entity (R11).\n\nAdjacently tagged on the wire — `{\"type\":\"server\"}`, `{\"type\":\"store\"}`, `{\"type\":\"machine\",\"id\":\"foo\"}` — because the `Machine` variant wraps a transparent-string newtype, which serde cannot internally tag.",
        "oneOf": [
          {
            "description": "The `codchi-server` daemon itself.",
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "server"
                ]
              }
            }
          },
          {
            "description": "The store container.",
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "store"
                ]
              }
            }
          },
          {
            "description": "A machine container.",
            "type": "object",
            "required": [
              "id",
              "type"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "machine"
                ]
              }
            }
          }
        ]
      },
...
      "MachineDetail": {
        "description": "The detail view of a machine.",
        "type": "object",
        "required": [
          "flake_lock_hash",
          "generations",
          "modules",
          "secrets",
          "view"
        ],
        "properties": {
          "flake_lock_hash": {
            "type": "string"
          },
          "generations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenerationView"
            }
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleSpec"
            }
          },
          "secrets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecretKey"
            }
          },
          "view": {
            "$ref": "#/components/schemas/MachineView"
          }
        }
      },
      "MachineView": {
        "description": "The list view of a machine. Returned from the daemon-maintained snapshot; read endpoints must not probe synchronously (P6).",
        "type": "object",
        "required": [
          "id",
          "run_status",
          "schema_version",
          "update_status"
        ],
        "properties": {
          "active_generation": {
            "description": "`None` ⇒ creating or failed-create (R9).",
            "default": null,
            "$ref": "#/components/schemas/GenerationId",
            "nullable": true
          },
          "busy_with": {
            "description": "The in-flight job owning this machine, if any.",
            "default": null,
            "$ref": "#/components/schemas/JobId",
            "nullable": true
          },
          "findings": {
            "description": "Active findings; `health = health(&findings)` (R10).",
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            }
          },
          "id": {
            "type": "string"
          },
          "last_reconcile_attempt_at": {
            "default": null,
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "last_reconciled_at": {
            "default": null,
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "run_status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "schema_version": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0.0
          },
          "snapshot_stale": {
            "default": false,
            "type": "boolean"
          },
          "update_status": {
            "$ref": "#/components/schemas/UpdateStatus"
          }
        }
      },
...
      "ModuleSpec": {
        "description": "A single module a machine pulls in. `url` is the codchi flake-url form (`<scheme>://<host>/<repo>?<query>#<attr>`, see beta `FlakeUrl`); `name` is the local handle used in CLI/config.",
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "properties": {
          "is_nixpkgs_source": {
            "description": "Whether this module supplies the machine's `nixpkgs` input (beta `nixpkgs_from`).",
            "default": false,
            "type": "boolean"
          },
          "name": {
            "description": "Local handle for the module (the map key in beta `MachineConfig`).",
            "type": "string"
          },
          "url": {
            "description": "Codchi flake-url including the module attribute path.",
            "type": "string"
          }
        }
      },
...
      "RunStatus": {
        "description": "Platform liveness, from the reconciler snapshot (P6).",
        "type": "string",
        "enum": [
          "stopped",
          "running"
        ]
      },
...
      "SecretKey": {
        "description": "A declared secret key. Values are never carried on this type.",
        "type": "object",
        "required": [
          "description",
          "has_value",
          "name"
        ],
        "properties": {
          "description": {
            "description": "Human-readable description from the NixOS declaration.",
            "type": "string"
          },
          "has_value": {
            "description": "Whether a value is currently stored for this key.",
            "type": "boolean"
          },
          "name": {
            "description": "Declared key name (matches the `codchi.secrets.env` attribute).",
            "type": "string"
          }
        }
      },
...
      "Severity": {
        "description": "Finding severity, ordered least → most severe.",
        "type": "string",
        "enum": [
          "info",
          "warning",
          "error",
          "critical"
        ]
      },
...
      "UpdateStatus": {
        "description": "Desired-vs-built config status (beta's `ConfigStatus`).",
        "type": "string",
        "enum": [
          "up_to_date",
          "needs_rebuild",
          "updates_available"
        ]
      },
...
    }
  }
}